refactor: DTO 위치 service layer로 이동 #173
