--- backend/src/main/java/ddangkong/controller/balance/room/dto/RoomJoinResponse.java
@@ -1,9 +1,0 @@
-package ddangkong.controller.balance.room.dto;
-
-import ddangkong.controller.balance.member.dto.MemberResponse;
-
-public record RoomJoinResponse(
-        Long roomId,
-        MemberResponse member
-) {
-}
