+++ backend/src/main/java/ddangkong/service/balance/room/dto/RoomInfoResponse.java
@@ -1,6 +1,6 @@
-package ddangkong.controller.balance.room.dto;
+package ddangkong.service.balance.room.dto;
 
-import ddangkong.controller.balance.member.dto.MemberResponse;
+import ddangkong.service.member.dto.MemberResponse;
 import ddangkong.domain.balance.room.Room;
 import ddangkong.domain.member.Member;
 import java.util.List;
