+++ backend/src/main/java/ddangkong/service/balance/room/dto/RoomJoinRequest.java
@@ -1,4 +1,4 @@
-package ddangkong.controller.balance.room.dto;
+package ddangkong.service.balance.room.dto;
 
 import jakarta.validation.constraints.NotBlank;
