+++ backend/src/main/java/ddangkong/service/balance/room/dto/RoomJoinResponse.java
@@ -1,0 +1,9 @@
+package ddangkong.service.balance.room.dto;
+
+import ddangkong.service.member.dto.MemberResponse;
+
+public record RoomJoinResponse(
+        Long roomId,
+        MemberResponse member
+) {
+}
