+++ backend/src/test/java/ddangkong/service/balance/vote/BalanceVoteServiceTest.java
@@ -7,13 +7,13 @@ import static ddangkong.support.fixture.MemberFixture.TACAN;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import ddangkong.controller.balance.content.dto.BalanceContentGroupResponse;
-import ddangkong.controller.balance.content.dto.BalanceContentTotalResponse;
-import ddangkong.controller.balance.option.dto.BalanceOptionGroupResponse;
-import ddangkong.controller.balance.option.dto.BalanceOptionTotalResponse;
-import ddangkong.controller.balance.vote.dto.BalanceVoteRequest;
-import ddangkong.controller.balance.vote.dto.BalanceVoteResponse;
-import ddangkong.controller.balance.vote.dto.BalanceVoteResultResponse;
+import ddangkong.service.balance.content.dto.BalanceContentGroupResponse;
+import ddangkong.service.balance.content.dto.BalanceContentTotalResponse;
+import ddangkong.service.balance.option.dto.BalanceOptionGroupResponse;
+import ddangkong.service.balance.option.dto.BalanceOptionTotalResponse;
+import ddangkong.service.balance.vote.dto.BalanceVoteRequest;
+import ddangkong.service.balance.vote.dto.BalanceVoteResponse;
+import ddangkong.service.balance.vote.dto.BalanceVoteResultResponse;
 import ddangkong.domain.balance.content.BalanceContent;
 import ddangkong.domain.balance.content.Category;
 import ddangkong.domain.balance.option.BalanceOption;
@@ -23,7 +23,7 @@ import ddangkong.domain.balance.vote.BalanceVote;
 import ddangkong.domain.member.Member;
 import ddangkong.exception.BadRequestException;
 import ddangkong.service.BaseServiceTest;
-import ddangkong.service.balance.vote.dto.VoteFinishedResponse;
+import ddangkong.service.balance.option.dto.VoteFinishedResponse;
 import ddangkong.support.annotation.FixedClock;
 import java.time.LocalDateTime;
 import java.util.List;
